flask folder

+++ dalle.py
@@ -30,37 +30,25 @@ def urlDownloader(url):
         out_file.write(data)
 
 
-
-
-
-
-#app = Flask(__name__)
+app = Flask(__name__)
 g = Github(os.environ["GITKEY"])
 
 
-#@app.route('/', methods=['GET', 'POST'])
-#def upload_image():
-    #if request.method == 'POST':
-imgurl = imgGenerator("I want a house that is beautiful for Cynthia to live in")
-response = requests.get(imgurl)
-file_path = "House.png"
-
-repo = g.get_repo("prajyaguru003/AIATL-2024")
-try:
-    contents = repo.get_contents(file_path, ref="main")
-    repo.delete_file(contents.path, "Delete image", contents.sha, branch="main")
-except Exception as e:
-    print("none exist")
-repo.create_file(file_path, "Add image", response.content, branch="main")
-
-        #return "Image uploaded successfully!"
-    #return '''
-    #<form method="post">
-    #  <input type="text" name="image_url" placeholder="Enter Image URL" />
-    #  <input type="submit" value="Upload">
-   # </form>
-    #'''
+@app.route('/add', methods=['GET', 'POST'])
+def upload_image():
+    if request.method == 'POST':
+        data = request.json
+        imgurl = imgGenerator(data['prompt'])
+        response = requests.get(imgurl)
+        file_path = "House.png"
+        repo = g.get_repo("prajyaguru003/AIATL-2024")
+        try:
+            contents = repo.get_contents(file_path, ref="main")
+            repo.delete_file(contents.path, "Delete image", contents.sha, branch="main")
+        except Exception:
+            print("none exist")
+        repo.create_file(file_path, "Add image", response.content, branch="main")
 
 
-#if __name__ == '__main__':
-#   app.run(debug=True)
+if __name__ == '__main__':
+    app.run(debug=True)
